First Commit
This folder contains my DSM501 dust sensor driver.
It is interrupt driven (using the PWM signals) hence is non-blocking.
It doesn't assume that the sample time is exactly 30s. It could be longer but not shorter.
It correctly handles the situation when sampling starts and the PWM signal is already low.
It correctly handles the situation when sampling ends and the PWM signal is still low.

Other implementations I have seen use pulseIn() to accumulate the PWM low time and assume the total run time is 30 seconds exactly.